requirements.txt
aenum==3.1.15
aiohappyeyeballs==2.6.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.3.0
APScheduler==3.11.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.4.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
Deprecated==1.2.14
deprecation==2.1.0
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
fonttools==4.52.4
frozenlist==1.4.1
future==1.0.0
google-api-core==2.19.1
google-api-python-client==2.140.0
google-auth==2.33.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gotrue==2.12.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
hyperframe==6.1.0
idna==3.7
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
line-bot-sdk==3.11.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
music==0.7b3
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.3
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
postgrest==1.0.1
propcache==0.3.1
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyJWT==2.10.1
pyparsing==3.1.2
PySocks==1.7.1
pytest==8.3.5
pytest-mock==3.14.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
realtime==2.4.2
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
schedule==1.2.2
scipy==1.13.1
selenium==4.31.0
setuptools==70.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spotipy==2.23.0
starlette==0.37.2
storage3==0.11.3
StrEnum==0.4.15
supabase==2.15.0
supafunc==0.9.4
sympy==1.12
termcolor==2.4.0
trio==0.29.0
trio-websocket==0.12.2
typer==0.12.3
typing_extensions==4.13.2
tzlocal==5.3.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.19.0
